refactor code and add check for transfer if both accounts belong to customer.
