+++ src/main/java/com/abc/bank/Account.java
@@ -1,9 +1,9 @@
-package com.abc;
+package com.abc.bank;
 
 import java.util.ArrayList;
 import java.util.List;
 
-import static com.abc.Utility.toDollars;
+import static com.abc.utility.Utility.toDollars;
 
 public abstract class Account {
     protected List<Transaction> transactions = new ArrayList<Transaction>();
@@ -11,22 +11,23 @@ public abstract class Account {
     protected double accountValue;
     private String accountStatement;
 
-    public Account(String accountType) {
+    Account(String accountType) {
         this.accountType = accountType;
         this.accountStatement = accountType + "\n";
     }
 
+    // Increases the account value and adds a transaction to the account.
     public void deposit(double amount) {
         if (amount <= 0) {
             throw new IllegalArgumentException("amount must be greater than zero");
         } else {
             transactions.add(new Transaction(amount));
             accountValue += amount;
-            accountStatement += "  deposit "
-                    + toDollars(amount) + "\n";
+            accountStatement += "  deposit " + toDollars(amount) + "\n";
         }
     }
 
+    // Decreases the account value and adds a transaction to the account.
     public void withdraw(double amount) {
         if (amount <= 0)
             throw new IllegalArgumentException("amount must be greater than zero");
