+++ src/main/java/com/abc/bank/Bank.java
@@ -1,4 +1,4 @@
-package com.abc;
+package com.abc.bank;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -22,8 +22,8 @@ public class Bank {
         return summary.toString();
     }
 
-    //Make sure correct plural of word is created based on the number passed in:
-    //If number passed in is 1 just return the word otherwise add an 's' at the end
+    // Make sure correct plural of word is created based on the number passed in:
+    // If number passed in is 1 just return the word otherwise add an 's' at the end
     private String handlePluralAccount(int number) {
         return number + " " + (number == 1 ? "account" : "accounts");
     }
@@ -31,7 +31,7 @@ public class Bank {
     public double totalInterestPaid() {
         double total = 0;
 
-        for(Customer customer : customers)
+        for (Customer customer : customers)
             total += customer.totalInterestEarned();
 
         return total;
