+++ src/main/java/com/abc/bank/MaxiSavingsAccount.java
@@ -1,4 +1,6 @@
-package com.abc;
+package com.abc.bank;
+
+import com.abc.utility.DateProvider;
 
 import java.util.Date;
