+++ src/main/java/com/abc/utility/DateProvider.java
@@ -1,4 +1,4 @@
-package com.abc;
+package com.abc.utility;
 
 import java.util.Calendar;
 import java.util.Date;
